requirements.txt
fastapi
uvicorn
pydantic
langchain_core
openai
numpy
scipy
dotenv
langchain_openai